src/requirements.txt
torch
torchvision
matplotlib
Pillow

